GPS订单历史轨迹

+++ jayud-oms/jayud-oms-model/src/main/java/com/jayud/oms/model/vo/PositionVO.java
@@ -124,6 +124,6 @@ public class PositionVO {
     private String vehicleStatus;
 
     @ApiModelProperty(value = "行驶里程")
-    private BigDecimal mile;
+    private Double mile;
 
 }
